requirements.txt
python-telegram-bot==20.6
requests==2.32.5
